Changed navbar

+++ js/main.js
@@ -20,6 +20,7 @@ var storeData = function (key){
     localStorage.setItem(key, JSON.stringify(userData));
     alert("Homework Added!");
     console.log(userData);
+    window.location.reload();
 };
 
 //Save data from form into local storage function
@@ -32,56 +33,41 @@ $('#view').on('pageinit', function(){
         displayData();
     });
 
+
 //Display data funciton
 var displayData = function(){
     $("#savedList").empty();
         for (var i= 0, j=localStorage.length; i<j ; i++){
             var key = localStorage.key(i);
             var item = JSON.parse(localStorage.getItem(key));
-            console.log(item);
+            //console.log(item);
             var makeSubList = $("<li></li>");
             var makeSubLi = $( "<h3>" + "Assignment "+item._id+"</h3>"+
-                "<p>"+item.fname+"</p>" +
+                "<p>"+item.datedue+"</p>" +
                 "<p>"+item.fname+"</p>" +
                 "<p>"+item.lname+"</p>"+
                 "<p>"+item.email+"</p>" +
                 "<p>"+item.subject+"</p>" +
                 "<p>"+item.notes+"</p>");
-            var makeLink = $("<a id='"+key+"'>Edit</a>");
+            var makeLink = $("<a id='"+key+"'>'Edit'</a>");
+            //console.log(item);
             makeLink.on('click', function(){
-               console.log(item);
+                    console.log(item);
             });
             makeLink.html(makeSubLi);
             makeSubList.append(makeLink).appendTo("#savedList");
             $("#savedList").listview("refresh");
         };
-};
+    };
 
-$('#view').on('click', function(){
+$('#view').on('pageinit', function(){
     $("#savedList").listview("refresh");
 });
-/*
-//Function to add json data 
-$('#loadjson').on('click', function(){
-    if(localStorage.length === 0){
-        $.ajax({
-                url      : "data.json",
-                type     : "GET",
-                dataType : "json",
-                success  : function(data, status) {
-                    console.log(status, data);
-                    displayData(data);
-                },
-                error   : function(error, parseerror) { 
-                    console.log(error, parseerror);
-                }
-            });
-        alert("Data has been added");
-    }
-});
-*/
+
 //Function to edit assignment
 var editAssignment = function(){
+
+        $("#savedList").empty();
         var value = localStorage.getItem(this.key);
         var userData = JSON.parse(value);
         
@@ -107,10 +93,46 @@ var editAssignment = function(){
         
     }
 
+//Clear Data
+    function boomData(){
+        if(localStorage.length === 0){
+            alert("There are no assignments.")
+        }else{
+            localStorage.clear();
+            alert("All homework has been deleted!");
+            window.location.reload();
+            return false;   
+        }
+    }
 
 //Clear data from local storage function
 $('#deleteAll').on('click', function(){
-  localStorage.clear();
+    boomData();
+});
+
+/*
+//Function to add json data 
+$('#loadjson').on('click', function(){
+    if(localStorage.length === 0){
+        $.ajax({
+                url      : "data.json",
+                type     : "GET",
+                dataType : "json",
+                success  : function(data, status) {
+                    console.log(status, data);
+                    displayData(data);
+                },
+                error   : function(error, parseerror) { 
+                    console.log(error, parseerror);
+                }
+            });
+        alert("Data has been added");
+    }
+});
+*/
+
+/*
+localStorage.clear();
   var confirmDelete = confirm("Are you sure you want to delete the assignment?");
     if(confirmDelete){
       localStorage.removeItem(this.key);
@@ -119,4 +141,4 @@ $('#deleteAll').on('click', function(){
     }else{
       alert("Assignment was NOT deleted!");
     }
-});
+*/
